Initial pass at using int64_t instead of long long
Even on systems without C99 where long long and stdint are both
missing, we can shim stdint and point it to any compiler-specific
type (i.e long long, _int64, etc.).

Also next is constant suffixes and determining what needs to
include stdint.

+++ src/hash/sha1/generic.h
@@ -11,7 +11,7 @@
 #include "hash/sha1.h"
 
 struct git_hash_sha1_ctx {
-	unsigned long long size;
+	uint64_t size;
 	unsigned int H[5];
 	unsigned int W[16];
 };
